requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
rich>=13.0.0
duckduckgo-search>=5.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0

